backend/requirements.txt
flask
flask_cors
tensorflow
Pillow
gunicorn